apps/api/pyproject.toml
[project]
name = "streaming-chat-api"
version = "0.1.0"
description = "Backend for the streaming chat workspace."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.16.5",
    "asyncpg>=0.30.0",
    "dbos>=0.20.0",
    "fastapi>=0.117.1",
    "httpx>=0.28.1",
    "orjson>=3.11.3",
    "pydantic-ai-slim[dbos,mcp,openai,temporal]>=1.18.0",
    "pydantic-settings>=2.11.0",
    "python-dotenv>=1.1.1",
    "redis>=6.4.0",
    "sqlalchemy>=2.0.43",
    "temporalio>=1.18.1",
    "uvicorn[standard]>=0.37.0",
]

[dependency-groups]
dev = [
    "aiosqlite>=0.21.0",
    "asgi-lifespan>=2.1.0",
    "fakeredis>=2.31.0",
    "pytest>=8.4.2",
    "pytest-asyncio>=1.2.0",
    "pytest-docker>=3.2.3",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/streaming_chat_api"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "-m 'not llm'"
markers = [
    "llm: tests that make real LLM/provider calls and only run when explicitly selected",
    "docker: tests that require dockerized services such as postgres",
]

[tool.ruff]
line-length = 100

[tool.ruff.format]
quote-style = "single"
docstring-code-format = false
